+++ CHANGELOG.md
@@ -10,12 +10,12 @@
 
 ## v0.0.4 [unreleased]
 
-- [ ] Update debug stuff (tiles counter)
+- [ ] Pause game if screen off, stop/start the press, stop/start the conveyor
+- [ ] ~Update debug stuff (tiles counter)~
 - [ ] Add some overlay (Pause and continue button)
 - [ ] Add Money system (overlay), get 100$ for a good tile and loose 200$ for a
       bad tile
 - [ ] Find a way to save the game state
-- [ ] Pause game if screen off, stop/start the press, stop/start the conveyor
 
 - Doing some code clean up
 - Adding `GameComponents` interface (`Input`, `BackGround`, `Engines`, `Conveyor`)